insurance_project/requirements.txt
django==4.2
faker
psycopg2-binary
django-admin-csvexport